✨ Function for send emails to employees
