+++ functions/src/send-reminders.ts
@@ -7,6 +7,7 @@ import { Config, isEnabled, asNumber } from "./config";
 import { daysBeforeCampaignEnds } from "./days-before-campaign-ends";
 import { enqueue } from "./process-email-queue";
 import { Employee } from "./import-employees-from-alibeez";
+import { sendEmailToEmployees } from "./send-email-vote";
 
 const db = firestore();
 const config = functions.config() as Config;
@@ -100,58 +101,7 @@ export const sendCampaignStartsReminder = functions.firestore
       return;
     }
 
-    const monthLongName = new Date(
-      Date.UTC(campaign.year, campaign.month)
-    ).toLocaleString("en-us", {
-      month: "long"
-    });
-
-    const employeeDocumentRefs = await db
-      .collection("employees")
-      .listDocuments();
-    for (const employeeDocumentRef of employeeDocumentRefs) {
-      const employeeDocument = await employeeDocumentRef.get();
-      const employee = employeeDocument.data() as Employee;
-      const message = {
-        from: config.features.reminders.voting_campaign_starts.sender,
-        to: employee.email,
-        subject: `Humeur du mois is open for ${monthLongName}!`,
-        html: `
-          <p>Hi ${employee.fullName},</p>
-          <p>
-            Tell us how it's been for you this past month!
-            Go to <a href="${linkToApp}">${linkToApp}</a>.
-          </p>
-          <p>See you soon!</p>
-          `,
-        "amp-html": `
-          <!doctype html>
-          <html ⚡4email data-css-strict>
-          <head>
-            <meta charset="utf-8">
-            <script async src="https://cdn.ampproject.org/v0.js"></script>
-            <style amp4email-boilerplate>body{visibility:hidden}</style>
-            <style amp-custom>
-              h1 {
-                margin: 1rem;
-              }
-            </style>
-          </head>
-          <body>
-            <p>Hi ${employee.fullName},</p>
-            <p>
-              Tell us how it's been for you this past month!
-              Go to <a href="${linkToApp}">${linkToApp}</a>.
-            </p>
-            <p>See you soon!</p>
-            <p>This email uses AMP!</p>
-          </body>
-          </html>
-        `
-      };
-
-      await enqueue(message);
-    }
+    sendEmailToEmployees(campaign)
 
     // await reminderRef.update({
     //   message
